Add files via upload
This project aims to predict house prices by applying advanced regression techniques on a comprehensive dataset of housing attributes. Utilizing a blend of Gradient Boosting and Random forest regressors, we seek to accurately forecast the price of a house based on its features, such as area, number of rooms, location, and more. This analysis provides valuable insights for real estate investors, buyers, and market analysts.